give the KS UI its own column
Instead of prepending to the right column, insert before the right
column. I decided to continue injecting into `#rightColumn` instead of
inserting before because getting a similar styling would be difficult
given that a lot of the rules (including themes) are attached to the
`#rightColumn` id.

I created a grid with many rows and made the UI span all of them. Each
element laid out after this moves into the second column and then
is laid out vertically for up to 20 rows. This gives some flexibility in
terms of new DOM nodes being added in the right column without needing
to adapt.

I had to strip some width-limiting rules so that there is sufficient
room for everything, but in doing so I realized the KS UI will attempt
to eat 100% of the entire width because of some `width: 100%` styles in
list-items which were previously necessary due to the heavy usage of
`float`. I stopped floating the list items and made it a `flow-root` to
avoid floats interacting with each other.